requirements.txt
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.1.0
matplotlib>=3.7.0
pyyaml>=6.0
openai>=1.0.0
python-dotenv>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.5.0
